% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxhash.R
\name{xxhash64}
\alias{xxhash64}
\alias{xxhash32}
\alias{xxhash128}
\alias{xxh3_64bits}
\title{Calculate hash of the data in a raw, numeric, integer or logical vector using \code{xxHash}}
\usage{
xxhash64(vec)

xxhash32(vec)

xxhash128(vec)

xxh3_64bits(vec)
}
\arguments{
\item{vec}{vector, matrix or array of raw, real, integer or logical values}
}
\description{
This function will calculate
the hash on the data payload within a raw, integer, numeric or logical vector.
It disregards dimensions and other attributes.
}
\details{
Vectors, matrices and arrays which contain the same data will hash to the same
value regardless of dimensions or attributes.
}
