% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxhash.R
\name{xxhash_file}
\alias{xxhash_file}
\title{Calculate the hash of a file}
\usage{
xxhash_file(file, algo = "xxh128")
}
\arguments{
\item{file}{filename}

\item{algo}{Select the specific xxhash algorithm. Default: 'xxh128'.
(the latest algorithm in the xxhash family)
Valid values: 'xxh32', 'xxh64', 'xxh128', 'xxh3'}
}
\description{
Calculate the hash of a file
}
