% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxhash.R
\name{xxhash_vec}
\alias{xxhash_vec}
\title{Calculate the hash of the numeric data within an atomic vector.}
\usage{
xxhash_vec(vec, algo = "xxh3_64bits")
}
\arguments{
\item{vec}{atomic vector (including matrices and arrays) containing raw,
integer, numeric, complex or logical values only.}

\item{algo}{Select the specific xxhash algorithm. Default: 'xxh3_64bits'.
(the latest algorithm in the xxhash family)
Valid values: 'xxhash32', 'xxhash64', 'xxhash128', 'xxh3_64bits'}
}
\description{
This function calculates the hash of only the data within an atomic vector, and
ignores all attributes (including dimensions).  By working in this way, it
avoids having to serialize the vector into raw bytes, and thus gains a speed
advantage.
}
\details{
The limitation of this method is that the same data in a vector will hash
to the same data stored in an array, even though the structure
of the data is different.

For general hashing of any R object, see
\code{\link[xxhashlite:xxhash]{xxhashlite::xxhash()}}
attributes and dimensions in hashing calculations.
}
