% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxhash.R
\name{xxhash}
\alias{xxhash}
\title{Calculate the hash of an arbitrary R object.}
\usage{
xxhash(robj, algo = "xxh3_64bits")
}
\arguments{
\item{robj}{Any R object}

\item{algo}{Select the specific xxhash algorithm. Default: 'xxh3_64bits'.
(the latest algorithm in the xxhash family)
Valid values: 'xxhash32', 'xxhash64', 'xxhash128', 'xxh3_64bits'}
}
\description{
This function will calculate the hash of any object understood by
\code{base::serialize()}.
}
\details{
If you are only interested in hashing the data contents of atomic vectors
containing numeric values, then
\code{\link[xxhashlite:xxhash_vec]{xxhashlite::xxhash_vec()}}
offers a speed advantage
over this function.
}
