% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxhash.R
\name{xxhash64}
\alias{xxhash64}
\alias{xxhash32}
\alias{xxhash128}
\alias{xxh3_64bits}
\title{Calculate hash of the given numeric, integer or logical vector using \code{xxHash}}
\usage{
xxhash64(vec)

xxhash32(vec)

xxhash128(vec)

xxh3_64bits(vec)
}
\arguments{
\item{vec}{vector of real, integer or logical values}
}
\description{
Vectors, matrices and arrays which contain the same data will hash to the same
value regardless of dimensions.
}
