% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxhash.R
\name{xxhash_raw}
\alias{xxhash_raw}
\title{Calculate the hash of a raw vector or string}
\usage{
xxhash_raw(robj, algo = "xxh3_64bits")
}
\arguments{
\item{robj}{raw vector or single character string}

\item{algo}{Select the specific xxhash algorithm. Default: 'xxh3_64bits'.
(the latest algorithm in the xxhash family)
Valid values: 'xxhash32', 'xxhash64', 'xxhash128', 'xxh3_64bits'}
}
\description{
Calculate the hash of a raw vector or string
}
